+++ src/app/app.ts
@@ -1,42 +1,43 @@
-import * as Sentry from '@sentry/node';
 import { SlashCreator, GatewayServer, SlashCommand, CommandContext } from 'slash-create';
 import Discord, { Client, ClientOptions, Intents, WSEventType } from 'discord.js';
 import path from 'path';
 import fs from 'fs';
-import Log, { LogUtils } from './utils/Log';
-import apiKeys from './service/constants/apiKeys';
-import constants from './service/constants/constants';
-import { RewriteFrames } from '@sentry/integrations';
+import { createLogger } from './utils/logger';
+import constants from './constants/constants';
+import NodeEventSource from 'eventsource';
+import axios from 'axios';
+// import { Cache } from './utils/cache';
 
-initializeSentryIO();
+// set api key header for axios globally
+axios.defaults.headers.common = {
+	'X-API-KEY': process.env.GOVERNATOR_API_KEY,
+};
+
+// initialize governator SSE
+const evtSource = new NodeEventSource(
+	constants.SSE_URL,
+	{ headers: { 'X-API-KEY': process.env.GOVERNATOR_API_KEY } },
+);
+
+// initialize cache
+// export const cache = new Cache();
+
+initializeGovernatorEvents();
+
+// initialize discordjs client and events
 const client: Client = initializeClient();
-initializeEvents();
 
+initializeDiscordJsEvents();
+
+// initialize slash-create
 const creator = new SlashCreator({
 	applicationID: process.env.DISCORD_BOT_APPLICATION_ID,
 	publicKey: process.env.DISCORD_BOT_PUBLIC_KEY,
 	token: process.env.DISCORD_BOT_TOKEN,
+	disableTimeouts: true,
 });
 
-creator.on('debug', (message) => Log.debug(`debug: ${ message }`));
-creator.on('warn', (message) => Log.warn(`warn: ${ message }`));
-creator.on('error', (error: Error) => Log.error(`error: ${ error }`));
-creator.on('synced', () => Log.debug('Commands synced!'));
-creator.on('commandRegister', (command: SlashCommand) => Log.debug(`Registered command ${command.commandName}`));
-creator.on('commandError', (command: SlashCommand, error: Error) => Log.error(`Command ${command.commandName}:`, {
-	indexMeta: true,
-	meta: {
-		name: error.name,
-		message: error.message,
-		stack: error.stack,
-		command,
-	},
-}));
-
-// Ran after the command has completed
-creator.on('commandRun', (command:SlashCommand, result: Promise<any>, ctx: CommandContext) => {
-	LogUtils.logCommandEnd(ctx);
-});
+initializeSlashCreateEvents();
 
 // Register command handlers
 creator
@@ -46,8 +47,9 @@ creator
 	.registerCommandsIn(path.join(__dirname, 'commands'))
 	.syncCommands();
 
-// Log client errors
-client.on('error', Log.error);
+
+// reload client
+client.destroy();
 
 client.login(process.env.DISCORD_BOT_TOKEN);
 
@@ -70,18 +72,24 @@ function initializeClient(): Client {
 	return new Discord.Client(clientOptions);
 }
 
-function initializeEvents(): void {
-	const eventFiles = fs.readdirSync(path.join(__dirname, '/events')).filter(file => file.endsWith('.js'));
+function initializeDiscordJsEvents(): void {
+	// get new winston child logger
+	const logger = createLogger('app:discordjs-events');
+	// Log client errors
+	client.on('error', (statement: string | any) => { logger.error(statement); });
+
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/events-discordjs')).filter(file => file.endsWith('.js'));
 	eventFiles.forEach(file => {
-		const event = new (require(`./events/${file}`).default)();
+		const event = new (require(`./events/events-discordjs/${file}`).default)();
 		try {
 			if (event.once) {
 				client.once(event.name, (...args) => event.execute(...args, client));
 			} else {
 				client.on(event.name, (...args) => event.execute(...args, client));
 			}
+			logger.debug(`Registered event ${event.name}`);
 		} catch (e) {
-			Log.error('Event failed to process', {
+			logger.error('Event failed to process', {
 				indexMeta: true,
 				meta: {
 					name: e.name,
@@ -94,18 +102,91 @@ function initializeEvents(): void {
 	});
 }
 
-function initializeSentryIO() {
-	Sentry.init({
-		dsn: `${apiKeys.sentryDSN}`,
-		tracesSampleRate: 1.0,
-		release: `${constants.APP_NAME}@${constants.APP_VERSION}`,
-		environment: `${process.env.SENTRY_ENVIRONMENT}`,
-		integrations: [
-			new RewriteFrames({
-				root: __dirname,
-			}),
-			new Sentry.Integrations.Http({ tracing: true }),
-		],
+function initializeSlashCreateEvents(): void {
+	// get new winston child logger
+	const logger = createLogger('app:slash-create-events');
+	creator.on('debug', (message) => logger.debug(message));
+	creator.on('warn', (message) => logger.warn(message));
+	creator.on('error', (error: Error) => logger.error(error));
+	creator.on('synced', () => logger.info('Commands synced!'));
+	// creator.on('commandRegister', (command: SlashCommand) => logger.info(`Registered command ${command.commandName}`));
+	creator.on('commandError', (command: SlashCommand, error: Error) => logger.error(`Command ${command.commandName}:`, {
+		indexMeta: true,
+		meta: {
+			name: error.name,
+			message: error.message,
+			stack: error.stack,
+			command,
+		},
+	}));
+	// Ran after the command has completed
+	creator.on('commandRun', (command:SlashCommand, result: Promise<any>, ctx: CommandContext) => {
+		logger.debug(`${ctx.user.username}#${ctx.user.discriminator} ran /${ctx.commandName}`, {
+			indexMeta: true,
+			meta: {
+				guildId: ctx.guildID,
+				userTag: `${ctx.user.username}#${ctx.user.discriminator}`,
+				userId: ctx.user.id,
+				params: ctx.options,
+			},
+		});
+	});
+
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/events-slash-create')).filter(file => file.endsWith('.js'));
+	eventFiles.forEach(file => {
+		const event = new (require(`./events/events-slash-create/${file}`).default)();
+		try {
+			if (event.once) {
+				creator.once(event.name, (...args) => event.execute(...args, client));
+			} else {
+				creator.on(event.name, (...args) => event.execute(...args, client));
+			}
+			logger.debug(`Registered event ${event.name}`);
+		} catch (e) {
+			logger.error('Slash-create event failed to process', {
+				indexMeta: true,
+				meta: {
+					name: e.name,
+					message: e.message,
+					stack: e.stack,
+					event,
+				},
+			});
+		}
+	});
+}
+
+function initializeGovernatorEvents(): void {
+	const logger = createLogger('app:governator-events');
+
+	evtSource.onerror = function(err) {
+		if (err) {
+			if (err.status === 401 || err.status === 403) {
+				logger.error('not authorized');
+			} else {
+				logger.error(err);
+			}
+		}
+	};
+
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/events-governator')).filter(file => file.endsWith('.js'));
+	eventFiles.forEach(file => {
+		const event = new (require(`./events/events-governator/${file}`).default)();
+		try {
+			evtSource.addEventListener(event.name, async (...args) => event.execute(...args, client));
+			logger.debug(`Registered event ${event.name}`);
+		} catch (e) {
+			logger.error(e);
+			logger.error('event failed to process', {
+				indexMeta: true,
+				meta: {
+					name: e.name,
+					message: e.message,
+					stack: e.stack,
+					event,
+				},
+			});
+		}
 	});
 }