+++ src/app/service/constants/channelIds.ts
@@ -1,4 +1,0 @@
-// TODO: add channel ID
-export default Object.freeze({
-	channelExample: 'channel-id-here',
-});